refactor: move Abort
